refactor: create MonthButtons for Header
