+++ src/features/monthButtons/index.js
@@ -1,0 +1,2 @@
+export {MonthButton} from './ui/MonthButton';
+export {useMonthButton} from './lib/useMonthButton';
