+++ src/features/monthButtons/ui/style.css
@@ -1,0 +1,23 @@
+.header__month-button {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 0.5rem;
+  background-color: transparent;
+  color: rgb(0, 0, 0);
+  border: none;
+  font-size: 1rem;
+  cursor: pointer;
+  padding: 0.5rem;
+  transition: all 0.5s;
+  border-radius: 50%;
+}
+
+.header__month-button--today {
+  border: 1px solid rgb(188, 188, 188);
+  border-radius: 0.25rem;
+}
+
+.header__month-button:hover {
+  background-color: rgba(236, 236, 236, 0.6);
+}
