+++ src/widgets/header/ui/style.css
@@ -18,33 +18,6 @@
   font-weight: 500;
 }
 
-.header__button {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 0.5rem;
-  background-color: transparent;
-  color: rgb(0, 0, 0);
-  border: none;
-  font-size: 1rem;
-  cursor: pointer;
-  padding: 0.5rem;
-  transition: all 0.5s;
-}
-
-.header__button:hover {
-  background-color: rgba(236, 236, 236, 0.6);
-}
-
-.header__button--today {
-  border: 1px solid rgb(188, 188, 188);
-  border-radius: 0.25rem;
-}
-
-.header__button--next, .header__button--prev {
-  border-radius: 50%;
-}
-
 .header__month {
   font-size: 1.5rem;
 }
